requirements.txt
azure-ai-evaluation
azure-identity
python-dotenv